requirements.txt
streamlit
pydub
SpeechRecognition
python-dotenv
google-generativeai